requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sqlalchemy>=1.4.0
apscheduler>=3.8.1
python-multipart>=0.0.5
email-validator>=1.1.3
alembic>=1.7.5